Move settings helpers to pretalx.common
